Bookings get deleted on payment. Customers get deleted once the last pending booking has been payed or cancled. Can now click on employee shedule to see details on the booking. Edit customer "save as customer" now shows what is int he database. Bookingform and its notification window fixed to show detail window with prober buttons. More inject statements for test data. Can't delete anything connected to a pending booking (except customer, then their bookings get deleted with them)

+++ 2nd.Semester.Eksamen.Infrastructure/Repositories/PersonRepositories/EmployeeRepositories/EmployeeRepository.cs
@@ -83,6 +83,8 @@ namespace _2nd.Semester.Eksamen.Infrastructure.Repositories.PersonRepositories.E
             using var transaction = await _context.Database.BeginTransactionAsync(System.Data.IsolationLevel.Serializable);
             try
             {
+                if (await _context.BookedTreatments.AnyAsync(t => t.EmployeeId == employee.Id))
+                    throw new Exception("Employee Cannot Be Deleted with pending treatments");
                 _context.Employees.Remove(employee);
                 await _context.SaveChangesAsync();
                 await transaction.CommitAsync();
